drivers: i2c: stm32: Move bus recovery sequence to common driver
Move the STM32 bus recovery sequence to STM32 I2C common driver
to factorize when used with the RTIO drivers flavors.
